move all classes to classes folder and rename files
