+++ sparrowpy/__init__.py
@@ -10,8 +10,8 @@ __version__ = '0.1.0'
 from . import geometry
 from . import radiosity
 from . import sound_object
-from .fast.radiosity_class import DirectionalRadiosityFast
-from .radiosity import (
+from .classes.RadiosityFast import DirectionalRadiosityFast
+from .classes.RadiosityKang import (
     RadiosityKang,
     DirectionalRadiosityKang,
     PatchesKang,
